refactor(features): extract single-flight setup scaffold
Dedupe the idempotent single-flight setup() lifecycle (AbortController
teardown) triplicated across notifications, presence, and push-notifications
into shared createSingleFlightSetup. Behavior unchanged.
